+++ lizard_shape/migrations/0001_initial.py
@@ -98,8 +98,8 @@ class Migration(SchemaMigration):
             ('min_value', self.gf('django.db.models.fields.CharField')(max_length=80, null=True, blank=True)),
             ('max_value', self.gf('django.db.models.fields.CharField')(max_length=80, null=True, blank=True)),
             ('is_exact', self.gf('django.db.models.fields.BooleanField')(default=False)),
-            ('color', self.gf('lizard_map.models.ColorField')(max_length=8, null=True, blank=True)),
-            ('color_inside', self.gf('lizard_map.models.ColorField')(max_length=8, null=True, blank=True)),
+            ('color', self.gf('lizard_map.fields.ColorField')(max_length=8, null=True, blank=True)),
+            ('color_inside', self.gf('lizard_map.fields.ColorField')(max_length=8, null=True, blank=True)),
             ('size', self.gf('django.db.models.fields.FloatField')(default=1.0)),
             ('icon', self.gf('django.db.models.fields.CharField')(max_length=80, null=True, blank=True)),
             ('mask', self.gf('django.db.models.fields.CharField')(max_length=80, null=True, blank=True)),
@@ -151,16 +151,16 @@ class Migration(SchemaMigration):
     models = {
         'lizard_map.legend': {
             'Meta': {'object_name': 'Legend'},
-            'default_color': ('lizard_map.models.ColorField', [], {'max_length': '8'}),
+            'default_color': ('lizard_map.fields.ColorField', [], {'max_length': '8'}),
             'descriptor': ('django.db.models.fields.CharField', [], {'max_length': '80'}),
             'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
-            'max_color': ('lizard_map.models.ColorField', [], {'max_length': '8'}),
+            'max_color': ('lizard_map.fields.ColorField', [], {'max_length': '8'}),
             'max_value': ('django.db.models.fields.FloatField', [], {'default': '100'}),
-            'min_color': ('lizard_map.models.ColorField', [], {'max_length': '8'}),
+            'min_color': ('lizard_map.fields.ColorField', [], {'max_length': '8'}),
             'min_value': ('django.db.models.fields.FloatField', [], {'default': '0'}),
             'steps': ('django.db.models.fields.IntegerField', [], {'default': '10'}),
-            'too_high_color': ('lizard_map.models.ColorField', [], {'max_length': '8'}),
-            'too_low_color': ('lizard_map.models.ColorField', [], {'max_length': '8'})
+            'too_high_color': ('lizard_map.fields.ColorField', [], {'max_length': '8'}),
+            'too_low_color': ('lizard_map.fields.ColorField', [], {'max_length': '8'})
         },
         'lizard_map.legendpoint': {
             'Meta': {'object_name': 'LegendPoint', '_ormbases': ['lizard_map.Legend']},
@@ -226,8 +226,8 @@ class Migration(SchemaMigration):
         },
         'lizard_shape.shapelegendsingleclass': {
             'Meta': {'object_name': 'ShapeLegendSingleClass'},
-            'color': ('lizard_map.models.ColorField', [], {'max_length': '8', 'null': 'True', 'blank': 'True'}),
-            'color_inside': ('lizard_map.models.ColorField', [], {'max_length': '8', 'null': 'True', 'blank': 'True'}),
+            'color': ('lizard_map.fields.ColorField', [], {'max_length': '8', 'null': 'True', 'blank': 'True'}),
+            'color_inside': ('lizard_map.fields.ColorField', [], {'max_length': '8', 'null': 'True', 'blank': 'True'}),
             'icon': ('django.db.models.fields.CharField', [], {'max_length': '80', 'null': 'True', 'blank': 'True'}),
             'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
             'is_exact': ('django.db.models.fields.BooleanField', [], {'default': 'False'}),